requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
pydantic<2.0
pydantic-core==2.18.2
httpx==0.24.0
pytest==7.4.0
pytest-asyncio==0.22.0
requests==2.31.0
python-multipart
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1
